+++ app/src/main/java/com/yaoh/AndroidDemo2/widgets/follow_view/OnAngleChangedListener.java
@@ -6,6 +6,6 @@ package com.yaoh.AndroidDemo2.widgets.follow_view;
 
 public interface OnAngleChangedListener {
 
-    public void onAngleChanged(float degree);
+    public void onAngleChanged(int degree);
 
 }
